Move BaseLegendItem type back to component
