+++ projects/js-packages/charts/src/components/legend/index.ts
@@ -1,4 +1,4 @@
 export { Legend } from './legend';
 export { useChartLegendItems } from './hooks/use-chart-legend-items';
-export type { LegendProps, BaseLegendProps } from './types';
+export type { LegendProps, BaseLegendProps, BaseLegendItem } from './types';
 export type { ChartLegendOptions } from './hooks/use-chart-legend-items';
