+++ projects/js-packages/charts/src/components/legend/types.ts
@@ -1,6 +1,6 @@
 import { LegendOrdinal } from '@visx/legend';
-import type { BaseLegendItem } from '../../types';
-import type { ComponentProps } from 'react';
+import { GlyphProps, LineStyles } from '@visx/xychart';
+import type { ComponentProps, CSSProperties, ReactNode } from 'react';
 
 // See https://airbnb.io/visx/docs/legend#Ordinal for more details.
 type LegendOrdinalProps = Omit< ComponentProps< typeof LegendOrdinal >, 'scale' | 'direction' >;
@@ -18,4 +18,17 @@ export type BaseLegendProps = Omit< LegendOrdinalProps, 'shapeStyle' > & {
 export type LegendProps = Omit< BaseLegendProps, 'items' > & {
 	items?: BaseLegendItem[];
 	chartId?: string;
+};
+
+export type BaseLegendItem = {
+	label: string;
+	value: number | string;
+	color: string;
+	glyphSize?: number;
+	renderGlyph?: < Datum extends object >( props: GlyphProps< Datum > ) => ReactNode;
+	shapeStyle?: CSSProperties & LineStyles;
+	// Optional group info for dynamic color resolution
+	group?: string;
+	index?: number;
+	overrideColor?: string;
 };
